move eps to module (#189)
